refactor(ios): decompose IosPeripheral facade from 411 to 286 lines via extension functions (#278)
Extract GATT operations to IosPeripheralGattOps.kt (135 lines):
- readGatt, writeGatt, observeGatt, observeValuesGatt
- readDescriptorGatt, writeDescriptorGatt
- readRssiGatt, requestMtuGatt

Extract ExperimentalBleApi stubs to IosPeripheralExtras.kt (38 lines):
- requestConnectionPriorityExt, requestConnectionParameterUpdateExt
- setPreferredPhyExt, readPhyExt

Extract restoration logic to IosPeripheralRestoration.kt (37 lines):
- restoreFromStateRestorationExt

Facade goes from 411 to 286 lines (125 line reduction, -30%).
Remove isochronous/subrating references from WIP decomposition files.
